ROFO-35 사용되지 않는 API 제거 및 약관 Response 삭제, 컬럼명 수정
클라이언트 작업자분과 의견을 나누어 다음과 같이 변경하였습니다.

스네이크 케이스 -> 카멜 케이스
required_flag -> isRequired
