+++ src/main/kotlin/kr/weit/roadyfoody/term/service/dto/DetailedTermResponse.kt
@@ -1,0 +1,21 @@
+package kr.weit.roadyfoody.term.service.dto
+
+import kr.weit.roadyfoody.term.domain.Term
+
+data class DetailedTermResponse(
+    val id: Long,
+    val title: String,
+    val isRequired: Boolean,
+    val content: String,
+) {
+    companion object {
+        @JvmStatic
+        fun from(term: Term): DetailedTermResponse =
+            DetailedTermResponse(
+                term.id,
+                term.title,
+                term.required,
+                term.content,
+            )
+    }
+}
