+++ src/main/kotlin/kr/weit/roadyfoody/term/service/dto/SummaryTermsResponse.kt
@@ -1,19 +1,14 @@
 package kr.weit.roadyfoody.term.service.dto
 
-import com.fasterxml.jackson.annotation.JsonProperty
 import kr.weit.roadyfoody.term.domain.Term
 
 data class SummaryTermsResponse(
-    @JsonProperty("all_terms_size")
     val allTermsSize: Int,
-    @JsonProperty("required_terms_size")
     val requiredTermsSize: Int,
-    @JsonProperty("optional_terms_size")
     val optionalTermsSize: Int,
     val terms: List<SummaryTermResponse>,
 ) {
     companion object {
-        @JvmStatic
         fun from(
             allTermsSize: Int,
             requiredTermsSize: Int,
@@ -31,15 +26,13 @@ data class SummaryTermsResponse(
 
 data class SummaryTermResponse(
     val title: String,
-    @JsonProperty("required_flag")
-    val requiredFlag: Boolean,
+    val isRequired: Boolean,
 ) {
     companion object {
-        @JvmStatic
         fun from(term: Term): SummaryTermResponse =
             SummaryTermResponse(
                 term.title,
-                term.requiredFlag,
+                term.required,
             )
     }
 }
